Fixed shift cards on volunteer dashboard
