+++ frontend/src/pages/VolunteerDash/index.js
@@ -5,7 +5,7 @@ import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import { DatePicker } from "@mui/x-date-pickers/DatePicker";
 import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
 import dayjs from "dayjs";
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useState, useCallback } from "react";
 import { getVolunteerShiftsForMonth } from "../../api/shiftService";
 import DashboardCoverage from "../../components/DashboardCoverage";
 import DashCoverShifts from "../../components/DashCoverShifts";
@@ -21,6 +21,7 @@ function VolunteerDash() {
   const [selectedDate, setSelectedDate] = useState(dayjs());
   const [future, setFuture] = useState(false);
 
+  // TODO: Change to become dynamic
   const checkInItem = () => {
     return checkIn ? (
       <div className="dash-check-in">
@@ -44,17 +45,34 @@ function VolunteerDash() {
     );
   };
 
-  useEffect(() => {
-    const fetchShifts = async () => {
-      const body = {
-        volunteer_id: volunteerID,
-        shiftDate: selectedDate.format("YYYY-MM-DD"),
-      };
-      const response = await getVolunteerShiftsForMonth(body);
-      setShifts(response);
+  const fetchShifts = useCallback(async () => {
+    const body = {
+      volunteer_id: volunteerID,
+      shiftDate: selectedDate.format("YYYY-MM-DD"),
     };
-    fetchShifts();
+    const response = await getVolunteerShiftsForMonth(body);
+
+    const shiftMap = new Map();
+    response.forEach((shift) => {
+      const existingShift = shiftMap.get(shift.shift_id);
+
+      if (existingShift && existingShift.shift_type === SHIFT_TYPES.MY_SHIFTS && shift.shift_type === SHIFT_TYPES.MY_COVERAGE_REQUESTS) {
+        shiftMap.set(shift.shift_id, shift);
+      } else if (!existingShift) {
+        shiftMap.set(shift.shift_id, shift);
+      }
+    })
+
+    setShifts(Array.from(shiftMap.values()));
+    
   }, [selectedDate, volunteerID]);
+
+  useEffect(() => {
+    const fetchData = async () => {
+        await fetchShifts();
+    }
+    fetchData();
+  }, [fetchShifts]);
 
   const allShifts = shifts.reduce((acc, shift) => {
     const date = shift.shift_date;
@@ -67,11 +85,12 @@ function VolunteerDash() {
 
   const groupedUpcomingShifts = shifts
     .filter((shift) => {
+      const shiftDay = dayjs(shift.shift_date).format('YYYY-MM-DD');
+      const shiftEnd = dayjs(`${shiftDay} ${shift.end_time}`);
+      const pastShift = dayjs().isAfter(shiftEnd);
+
       return (
-        shift.shift_type ===
-          (SHIFT_TYPES.MY_SHIFTS || SHIFT_TYPES.MY_COVERAGE_REQUESTS) &&
-        dayjs(shift.shift_date).isAfter(monthDate) &&
-        dayjs(shift.shift_date).isAfter(dayjs())
+        (shift.shift_type === SHIFT_TYPES.MY_SHIFTS || shift.shift_type === SHIFT_TYPES.MY_COVERAGE_REQUESTS) && !pastShift
       );
     })
     .reduce((acc, shift) => {
@@ -97,14 +116,6 @@ function VolunteerDash() {
     }, {});
 
   const handleShiftUpdate = () => {
-    const fetchShifts = async () => {
-      const body = {
-        volunteer_id: volunteerID,
-        shiftDate: selectedDate.format("YYYY-MM-DD"),
-      };
-      const response = await getVolunteerShiftsForMonth(body);
-      setShifts(response);
-    };
     fetchShifts();
   };
 
@@ -125,7 +136,7 @@ function VolunteerDash() {
                 color: "var(--primary-blue)",
               }}
               value={selectedDate}
-              onChange={(newValue) => setSelectedDate(newValue.day(2))}
+              onChange={(newValue) => setSelectedDate(newValue.date(1))}
             />
           </LocalizationProvider>
         </div>
@@ -159,6 +170,7 @@ function VolunteerDash() {
             groupedShifts={future ? groupedUpcomingShifts : allShifts}
             future={future}
             handleShiftUpdate={handleShiftUpdate}
+            volunteerID={volunteerID}
           />
         </div>
 
@@ -168,6 +180,7 @@ function VolunteerDash() {
               future={future}
               groupedShifts={groupedCoverShifts}
               handleShiftUpdate={handleShiftUpdate}
+              volunteerID={volunteerID}
             />
           </div>
           {checkInItem()}
